moved styles object to folder, cleaned up unused imports
