+++ screens/FiveByFive/FiveByFive.js
@@ -7,15 +7,15 @@ import {
   Text,
   Platform,
 } from 'react-native';
-const image = require('../landscapeSmall.png');
-import roundDots from '../Helpers/roundDots';
-const RapidBreathing = ({ route, navigation }) => {
+import renderStyle from './FiveByFiveStyles';
+import roundDots from '../../Helpers/roundDots';
+const FiveByFive = ({ route, navigation }) => {
   const { numberOfCycles } = route.params;
   const [cycle, setCycle] = useState(0);
-  const breathingCircle = useRef(new Animated.Value(1)).current;
-  const outerCircle = useRef(new Animated.Value(1)).current;
   const [displayText, setDisplayText] = useState('In');
   const [animationEnabled, setAnimationEnabled] = useState(false);
+
+  const [styles, breathingCircle, outerCircle] = renderStyle();
   useEffect(() => {
     if (!animationEnabled) {
       setAnimationEnabled(true);
@@ -38,74 +38,18 @@ const RapidBreathing = ({ route, navigation }) => {
       setAnimationEnabled(false);
     };
   });
-  const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      alignContent: 'center',
-      alignItems: 'center',
-      justifyContent: 'center',
-      backgroundColor: 'gray',
-    },
-    ball: {
-      height: 70,
-      width: 70,
-      borderRadius: 40,
-      borderWidth: 40,
-      borderColor: 'rgba(255, 255, 255, 0.85)',
-      transform: [{ scale: breathingCircle }],
-      // top: 10,
-    },
-    round_container_5: {
-      flexDirection: 'row',
-      justifyContent: 'center',
-      top: 18,
-    },
-    dot_5: {
-      height: 16,
-      width: 16,
-      marginRight: 4,
-      marginLeft: 4,
-      borderRadius: 50,
-      borderStyle: 'solid',
-      borderWidth: 2,
-      borderColor: '#7F6C72',
-    },
-    gradientCircle: {
-      flex: 1,
-      height: 140,
-      width: 140,
-      borderRadius: 100,
-      position: 'absolute',
-      top: Platform.OS === 'android' ? 210 : 280,
-      transform: [{ scale: outerCircle }],
-      backgroundColor: 'rgba(255, 255, 255, 0.4)',
-    },
-    text: {
-      alignSelf: 'center',
-      bottom: 60,
-      fontFamily: 'Lato-Bold',
-      fontSize: 30,
-      position: 'relative',
-    },
-    currentRound: {
-      fontFamily: 'Lato-Bold',
-      color: 'white',
-      fontSize: 44,
-      position: 'absolute',
-    },
-  });
 
   const TimerText = () => {
     Animated.parallel([
       Animated.timing(breathingCircle, {
         toValue: 8,
-        duration: 2000,
+        duration: 5500,
         useNativeDriver: true,
         Easing: Easing.bezier(0.65, 0, 0.35, 1),
       }),
       Animated.timing(outerCircle, {
         toValue: 8,
-        duration: 2000,
+        duration: 5500,
         useNativeDriver: true,
         Easing: Easing.bezier(0.65, 0, 0.35, 1),
       }),
@@ -117,13 +61,13 @@ const RapidBreathing = ({ route, navigation }) => {
       Animated.parallel([
         Animated.timing(breathingCircle, {
           toValue: 1,
-          duration: 1300,
+          duration: 5500,
           useNativeDriver: true,
           Easing: Easing.bezier(0.65, 0, 0.35, 1),
         }),
         Animated.timing(outerCircle, {
           toValue: 1,
-          duration: 1300,
+          duration: 5500,
           useNativeDriver: true,
           Easing: Easing.bezier(0.65, 0, 0.35, 1),
         }),
@@ -160,6 +104,7 @@ const RapidBreathing = ({ route, navigation }) => {
             flexDirection: 'column-reverse',
             justifyContent: 'flex-end',
             height: 8,
+            // width: '5%',
             bottom: 44,
             right: 10,
           }}>
@@ -170,4 +115,4 @@ const RapidBreathing = ({ route, navigation }) => {
   );
 };
 
-export default RapidBreathing;
+export default FiveByFive;
