+++ screens/rapidBreathing/RapidBreathing.js
@@ -1,19 +1,11 @@
-import React, { useEffect, useRef, useState } from 'react';
-import {
-  Animated,
-  Easing,
-  View,
-  StyleSheet,
-  Text,
-  Platform,
-} from 'react-native';
-const image = require('../landscapeSmall.png');
-import roundDots from '../Helpers/roundDots';
-const FiveByFive = ({ route, navigation }) => {
+import React, { useEffect, useState } from 'react';
+import { Animated, Easing, View, Text } from 'react-native';
+import renderStyle from './RapidBreathingStyles';
+import roundDots from '../../Helpers/roundDots';
+const RapidBreathing = ({ route, navigation }) => {
   const { numberOfCycles } = route.params;
   const [cycle, setCycle] = useState(0);
-  const breathingCircle = useRef(new Animated.Value(1)).current;
-  const outerCircle = useRef(new Animated.Value(1)).current;
+  const [styles, breathingCircle, outerCircle] = renderStyle();
   const [displayText, setDisplayText] = useState('In');
   const [animationEnabled, setAnimationEnabled] = useState(false);
   useEffect(() => {
@@ -38,76 +30,18 @@ const FiveByFive = ({ route, navigation }) => {
       setAnimationEnabled(false);
     };
   });
-  const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      alignContent: 'center',
-      alignItems: 'center',
-      justifyContent: 'center',
-      backgroundColor: 'black',
-    },
-    ball: {
-      height: 70,
-      width: 70,
-      borderRadius: 40,
-      borderWidth: 40,
-      borderColor: 'rgba(255, 255, 255, 0.85)',
-      transform: [{ scale: breathingCircle }],
-    },
-    round_container_5: {
-      flexDirection: 'row',
-      justifyContent: 'center',
-      // marginTop: 10,
-      top: 18,
-      // zIndex: 1,
-    },
-    dot_5: {
-      height: 16,
-      width: 16,
-      marginRight: 4,
-      marginLeft: 4,
-      borderRadius: 50,
-      borderStyle: 'solid',
-      borderWidth: 2,
-      borderColor: '#7F6C72',
-      backgroundColor: cycle > 1 ? '#7F6C72' : 'black',
-    },
-    gradientCircle: {
-      flex: 1,
-      height: 140,
-      width: 140,
-      borderRadius: 100,
-      position: 'absolute',
-      top: Platform.OS === 'android' ? 210 : 280,
-      transform: [{ scale: outerCircle }],
-      backgroundColor: 'rgba(255, 255, 255, 0.4)',
-    },
-    text: {
-      alignSelf: 'center',
-      bottom: 60,
-      fontFamily: 'Lato-Bold',
-      fontSize: 30,
-      position: 'relative',
-    },
-    currentRound: {
-      fontFamily: 'Lato-Bold',
-      color: 'white',
-      fontSize: 44,
-      position: 'absolute',
-    },
-  });
 
   const TimerText = () => {
     Animated.parallel([
       Animated.timing(breathingCircle, {
         toValue: 8,
-        duration: 5500,
+        duration: 2000,
         useNativeDriver: true,
         Easing: Easing.bezier(0.65, 0, 0.35, 1),
       }),
       Animated.timing(outerCircle, {
         toValue: 8,
-        duration: 5500,
+        duration: 2000,
         useNativeDriver: true,
         Easing: Easing.bezier(0.65, 0, 0.35, 1),
       }),
@@ -119,13 +53,13 @@ const FiveByFive = ({ route, navigation }) => {
       Animated.parallel([
         Animated.timing(breathingCircle, {
           toValue: 1,
-          duration: 5500,
+          duration: 1300,
           useNativeDriver: true,
           Easing: Easing.bezier(0.65, 0, 0.35, 1),
         }),
         Animated.timing(outerCircle, {
           toValue: 1,
-          duration: 5500,
+          duration: 1300,
           useNativeDriver: true,
           Easing: Easing.bezier(0.65, 0, 0.35, 1),
         }),
@@ -162,7 +96,6 @@ const FiveByFive = ({ route, navigation }) => {
             flexDirection: 'column-reverse',
             justifyContent: 'flex-end',
             height: 8,
-            // width: '5%',
             bottom: 44,
             right: 10,
           }}>
@@ -173,4 +106,4 @@ const FiveByFive = ({ route, navigation }) => {
   );
 };
 
-export default FiveByFive;
+export default RapidBreathing;
